fix test

+++ src/main/java/uk/gov/hmcts/reform/em/hrs/service/email/HearingReportCsvWriter.java
@@ -44,7 +44,7 @@ public class HearingReportCsvWriter {
                     hearingRecSeg.getIngestionFileSourceUri(),
                     hearingRecSeg.getHearingRecording().getHearingSource(),
                     hearingRecSeg.getHearingRecording().getServiceCode(),
-                    Math.ceil((float) hearingRecSeg.getFileSizeMb() / 1000),
+                    (int)Math.ceil((float) hearingRecSeg.getFileSizeMb() / 1000),
                     hearingRecSeg.getHearingRecording().getCcdCaseId(),
                     hearingRecSeg.getCreatedOn()
                 );
